requirements.txt
PyMuPDF==1.23.20
pillow==10.2.0
tabula-py==2.9.0
pandas==2.2.0
opencv-python==4.9.0.80
keras-ocr==0.9.3
tensorflow==2.15.0
requests==2.31.0
python-dotenv==1.0.1